\name{contingencyCcoef}
\alias{contingencyCcoef}
\title{Caluculate Contingency coefficient C of cross table}
\arguments{
  \item{quga}{QuGAcomp object}
}
\description{
  This function caluculate Contingency coefficient C of
  cross table (contingency table).
}
\examples{
data(quga)

quga.contingencyC <- contingencyCcoef(quga)
}

