\name{pearsonCoef}
\alias{pearsonCoef}
\title{Caluculate Pearson product-moment correlation coefficient}
\arguments{
  \item{quga.ctable}{QuGAcomp CrossTable object}
}
\description{
  This function caluculate Pearson product-moment
  correlation coefficient
}
\examples{
data(quga)

quga.pearson <- pearsonCoef(quga)
}

