\docType{package}
\name{QuGAcomp-package}
\alias{package-qugacomp}
\alias{qugacomp-package}
\alias{QuGAcomp-package}
\title{Quantitative Genome Annotation Comparison Tool}
\description{
  This R package can quantitative compare genome
  annotations such as peaks of two ChIP-seq data.
}
\references{
  Itoshi NIAIDO and Hiroki R. Ueda., QuGAcpmp: Quantitative
  genome annoation comparison Tool for high-throughput
  sequencing data. (in prep.)
}

