\docType{data}
\name{sox2.gr}
\alias{sox2.gr}
\title{GenomicRanges object of Sox2 ChIP-seq data}
\format{GenomicRanges object}
\source{
  \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE11431}
}
\description{
  This object is Sox2 ChIP-seq data in GenomicRanges object
}
\references{
  Xi Chen, et. al. Integration of External Signaling
  Pathways with the Core Transcriptional Network in
  Embryonic Stem Cells, Cell, Volume 133, Issue 6, 13 June
  2008, Pages 1106-1117, ISSN 0092-8674,
  10.1016/j.cell.2008.04.043.
}
\keyword{datasets}

